shellcheck in normal mode for *.t files in tests/

+++ tests/registration/00-simple.t
@@ -32,10 +32,10 @@ init_suite "${TEST_NAME_BASE}" <<'__SUITE_RC__'
 __SUITE_RC__
 
 # Unique suite run-dir prefix to avoid messing with real suites.
-PRE=cylctb-reg-${CYLC_TEST_TIME_INIT}
+PRE="cylctb-reg-${CYLC_TEST_TIME_INIT}"
 
 # Test fail no suite.rc file.
-CYLC_RUN_DIR=$(cylc get-global --print-run-dir)
+CYLC_RUN_DIR="$(cylc get-global --print-run-dir)"
 TEST_NAME="${TEST_NAME_BASE}-noreg"
 run_fail "${TEST_NAME}" cylc register "${SUITE_NAME}" "${PWD}/zilch"
 contains_ok "${TEST_NAME}.stderr" <<__ERR__
@@ -45,35 +45,35 @@ __ERR__
 CHEESE=${PRE}-cheese
 # Test default name: "cylc reg" (suite in $PWD, no args)
 TEST_NAME="${TEST_NAME_BASE}-cheese"
-mkdir $CHEESE
-cd $CHEESE
-touch suite.rc
+mkdir "${CHEESE}"
+cd "${CHEESE}" || exit 1
+touch 'suite.rc'
 run_ok "${TEST_NAME}" cylc register
 contains_ok "${TEST_NAME}.stdout" <<__OUT__
-REGISTERED $CHEESE -> ${PWD}
+REGISTERED ${CHEESE} -> ${PWD}
 __OUT__
-cd ..
-rm -rf "${CYLC_RUN_DIR}/$CHEESE"
+cd .. || exit 1
+rm -rf "${CYLC_RUN_DIR:?}/${CHEESE}"
 
 # Test default name: "cylc reg REG" (suite in $PWD)
 TEST_NAME="${TEST_NAME_BASE}-toast"
-cd $CHEESE
-TOAST=${PRE}-toast
-run_ok "${TEST_NAME}" cylc register $TOAST
+cd "${CHEESE}" || exit 1
+TOAST="${PRE}-toast"
+run_ok "${TEST_NAME}" cylc register "${TOAST}"
 contains_ok "${TEST_NAME}.stdout" <<__OUT__
-REGISTERED $TOAST -> ${PWD}
+REGISTERED ${TOAST} -> ${PWD}
 __OUT__
-cd ..
-rm -rf "${CYLC_RUN_DIR}/$TOAST"
+cd .. || exit 1
+rm -rf "${CYLC_RUN_DIR:?}/$TOAST"
 
 # Test "cylc reg REG PATH"
 TEST_NAME="${TEST_NAME_BASE}-bagels"
-BAGELS=${PRE}-bagels
-run_ok "${TEST_NAME}" cylc register $BAGELS $CHEESE
+BAGELS="${PRE}-bagels"
+run_ok "${TEST_NAME}" cylc register "${BAGELS}" "${CHEESE}"
 contains_ok "${TEST_NAME}.stdout" <<__OUT__
-REGISTERED $BAGELS -> ${PWD}/$CHEESE
+REGISTERED ${BAGELS} -> ${PWD}/${CHEESE}
 __OUT__
-rm -rf "${CYLC_RUN_DIR}/$BAGELS"
+rm -rf "${CYLC_RUN_DIR:?}/${BAGELS}"
 
 # Test "cylc reg REG ~/cylc-run/REG"
 TEST_NAME="${TEST_NAME_BASE}-onion"
@@ -93,40 +93,41 @@ REGISTERED ${ONION} -> ${CYLC_RUN_DIR}/${ONION}
 __OUT__
 SOURCE="$(readlink "${CYLC_RUN_DIR}/${ONION}/.service/source")"
 run_ok "${TEST_NAME}-2-source" test '..' = "${SOURCE}"
-rm -rf "${CYLC_RUN_DIR}/${ONION}"
+rm -rf "${CYLC_RUN_DIR:?}/${ONION}"
 
 # Test fail "cylc reg REG PATH" where REG already points to PATH2
-YOGHURT=${PRE}-YOGHURT
-cp -r $CHEESE $YOGHURT
+YOGHURT="${PRE}-YOGHURT"
+cp -r "${CHEESE}" "${YOGHURT}"
 TEST_NAME="${TEST_NAME_BASE}-cheese"
-run_ok "${TEST_NAME}" cylc register $CHEESE $CHEESE
+run_ok "${TEST_NAME}" cylc register "${CHEESE}" "${CHEESE}"
 TEST_NAME="${TEST_NAME_BASE}-repurpose1"
-run_fail "${TEST_NAME}" cylc register $CHEESE $YOGHURT
+run_fail "${TEST_NAME}" cylc register "${CHEESE}" "${YOGHURT}"
 contains_ok "${TEST_NAME}.stderr" <<__ERR__
-SuiteServiceFileError: the name '$CHEESE' already points to ${PWD}/$CHEESE.
+SuiteServiceFileError: the name '${CHEESE}' already points to ${PWD}/${CHEESE}.
 Use --redirect to re-use an existing name and run directory.
 __ERR__
 
 # Test succeed "cylc reg REG PATH" where REG already points to PATH2
 TEST_NAME="${TEST_NAME_BASE}-repurpose2"
-cp -r $CHEESE $YOGHURT
-run_ok "${TEST_NAME}" cylc register --redirect $CHEESE $YOGHURT
+cp -r "${CHEESE}" "${YOGHURT}"
+run_ok "${TEST_NAME}" cylc register --redirect "${CHEESE}" "${YOGHURT}"
 sed -i 's/^\t//; s/^.* WARNING - /WARNING - /' "${TEST_NAME}.stderr"
 contains_ok "${TEST_NAME}.stderr" <<__ERR__
-WARNING - the name '$CHEESE' points to ${PWD}/$CHEESE.
-It will now be redirected to ${PWD}/$YOGHURT.
-Files in the existing $CHEESE run directory will be overwritten.
+WARNING - the name '${CHEESE}' points to ${PWD}/${CHEESE}.
+It will now be redirected to ${PWD}/${YOGHURT}.
+Files in the existing ${CHEESE} run directory will be overwritten.
 __ERR__
 contains_ok "${TEST_NAME}.stdout" <<__OUT__
-REGISTERED $CHEESE -> ${PWD}/$YOGHURT
+REGISTERED ${CHEESE} -> ${PWD}/${YOGHURT}
 __OUT__
-rm -rf "${CYLC_RUN_DIR}/$CHEESE"
+rm -rf "${CYLC_RUN_DIR:?}/${CHEESE}"
 
 run_ok "${TEST_NAME_BASE}-get-dir" cylc get-directory "${SUITE_NAME}"
 
-cd .. # necessary so the suite is being validated via the database not filepath
+# necessary so the suite is being validated via the database not filepath
+cd .. || exit 1
 run_ok "${TEST_NAME_BASE}-val" cylc validate "${SUITE_NAME}"
-cd "${OLDPWD}"
+cd "${OLDPWD}" || exit 1
 
 run_ok "${TEST_NAME_BASE}-print" cylc print
 contains_ok "${TEST_NAME_BASE}-print.stdout" <<__OUT__
@@ -135,9 +136,9 @@ __OUT__
 
 # Filter out errors from 'bad' suites in the 'cylc-run' directory
 NONSPECIFIC_ERR2='\[Errno 2\] No such file or directory:'
-SPECIFIC_ERR2="$NONSPECIFIC_ERR2 '$HOME/cylc-run/$SUITE_NAME/suite.rc'"
-ERR2_COUNT=$(grep -c "$SPECIFIC_ERR2" "${TEST_NAME_BASE}-print.stderr")
-if [ "$ERR2_COUNT" -eq "0" ]; then
+SPECIFIC_ERR2="$NONSPECIFIC_ERR2 '$HOME/cylc-run/${SUITE_NAME}/suite.rc'"
+ERR2_COUNT="$(grep -c "$SPECIFIC_ERR2" "${TEST_NAME_BASE}-print.stderr")"
+if ((ERR2_COUNT == 0)); then
     grep -v -s "$NONSPECIFIC_ERR2" "${TEST_NAME_BASE}-print.stderr" > "${TEST_NAME_BASE}-print-filtered.stderr"
     cmp_ok "${TEST_NAME_BASE}-print-filtered.stderr" <'/dev/null'
 else
